Create 1012_improved.c
This problem has really taught me a trick. You have to store the answer you got before for repeated input. In the past, I thought this is useless, for I thought we seldom meet this situation in which repeated input is previded. In fact, it is common. I still have a lot to learn.